useLogout, useToken custom hook 작성 후 App.js의 유저 정보 조회 코드에 적용 #298
